4. Перенос методов в Stream API [#279131 #424830]

+++ src/main/java/ru/job4j/stream/ProductLabel.java
@@ -6,7 +6,7 @@ public class ProductLabel {
     public List<String> generateLabels(List<Product> products) {
         return products.stream().filter(product -> product.getStandard() - product.getActual() < 4)
                 .filter(product -> product.getStandard() - product.getActual() >= 0)
-                .map(product -> new Label(product.getName(), product.getPrice() * (float) 0.5))
+                .map(product -> new Label(product.getName(), product.getPrice() * 0.5f))
                 .map(Label::toString)
                 .toList();
